requirements.txt
Flask==3.0.3
langchain>=1.0.0
langchain-community>=0.3.0
langchain-core>=1.0.0
langchain-text-splitters>=0.3.0
openai>=1.0
pinecone-client[grpc]==3.2.1
python-dotenv>=1.0
langchain-openai>=0.2.0
pypdf==6.0.0
flask-cors==6.0.1
